+++ src/assets/demo/loadingButton.js
@@ -1,5 +1,7 @@
 // example for adding and removing loading state from button
 function loadingButton(el, duration) {
   el.setAttribute("loading", true);
-  setTimeout(() => el.setAttribute("loading", false), duration);
+  setTimeout(function() {
+    el.setAttribute("loading", false);
+  }, duration);
 }