+++ src/components/calcite-button/calcite-button.tsx
@@ -30,6 +30,9 @@ export class CalciteButton {
     | "clear"
     | "inline" = "solid";
 
+  /** Select theme (light or dark) */
+  @Prop({ reflect: true }) theme: "light" | "dark" = "light";
+
   /** specify the scale of the button, defaults to m */
   @Prop({ mutable: true, reflect: true }) scale: "xs" | "s" | "m" | "l" | "xl" =
     "m";
@@ -66,6 +69,9 @@ export class CalciteButton {
 
     let width = ["auto", "half", "full"];
     if (!width.includes(this.width)) this.width = "auto";
+
+    let theme = ["dark", "light"];
+    if (!theme.includes(this.theme)) this.theme = "light";
   }
 
   componentDidLoad() {