+++ src/components/calcite-radio-group-item/calcite-radio-group-item.tsx
@@ -70,6 +70,9 @@ export class CalciteRadioGroupItem {
     }
 
     this.inputProxy = inputProxy;
+
+    const futureSlotted = Array.from(this.el.childNodes);
+    this.hasLabel = futureSlotted.some((child) => child.nodeType === Node.TEXT_NODE);
   }
 
   disconnectedCallback() {
@@ -82,7 +85,7 @@ export class CalciteRadioGroupItem {
     return (
       <Host role="radio" aria-checked={checked ? "true" : "false"}>
         <label>
-          <slot>{value}</slot>
+          {this.hasLabel ? <slot /> : value}
           <slot name="input" />
         </label>
       </Host>
@@ -103,6 +106,8 @@ export class CalciteRadioGroupItem {
   //  Private State/Props
   //
   //--------------------------------------------------------------------------
+
+  private hasLabel = false;
 
   private inputProxy: HTMLInputElement;