+++ src/interfaces/TreeSelectionMode.ts
@@ -1,0 +1,6 @@
+export enum TreeSelectionMode {
+  Single = "single",
+  Multi = "multi",
+  Children = "children",
+  MultiChildren = "multi-children"
+}