+++ CHANGELOG.md
@@ -7,6 +7,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### [Unreleased][HEAD]
 
+### Added
+- dark theme for `calcite-slider`
+
+### Fixed
+- solved issue with incorrect positioning of handle in `calcite-slider`
+
+## [v1.0.0-beta.3] - Aug 16th 2019
+
+### Added
+- date picker keyboard support
+- date picker page-up and page-down buttons
+- pre-render support for existing components
+
+### Fixes
+- style updates/dark theme for buttons
+- fixed styling of modals in firefox
+- fixed radio-group styling in Edge
+- pointed calcite-base to correct npm version
+
 ## [v1.0.0-beta.2] - Aug 2nd 2019
 
 Fix issue with previous release.
@@ -15,6 +34,7 @@ Fix issue with previous release.
 
 First initial beta release.
 
+[v1.0.0-beta.3]: https://github.com/ArcGIS/calcite-components/compare/v1.0.0-beta.2...v1.0.0-beta.3 "v1.0.0-beta.3"
 [v1.0.0-beta.2]: https://github.com/ArcGIS/calcite-components/compare/v1.0.0-beta.1...v1.0.0-beta.2 "v1.0.0-beta.2"
 [v1.0.0-beta.1]: https://github.com/ArcGIS/calcite-components/compare/dafb2312835ec6fef134d0d2b20aabd1dfe907cf...v1.0.0-beta.1 "v1.0.0-beta.1"
-[HEAD]: https://github.com/ArcGIS/calcite-components/compare/v1.0.0-beta.1...HEAD "Unreleased Changes"
+[HEAD]: https://github.com/ArcGIS/calcite-components/compare/v1.0.0-beta.3...HEAD "Unreleased Changes"